requirements.txt
asgiref==3.3.4
Django==3.2.2
django-advanced-filters==1.3.0
django-bootstrap-form==3.4
django-braces==1.14.0
django-filter==2.4.0
pkg-resources==0.0.0
psycopg2-binary==2.8.6
pytz==2021.1
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
